Prepare for pull
